#56 ユーザープロフィールのコンポーネントを作成
